Fikset mult utregning for slots, mm

+++ player.h
@@ -4,16 +4,16 @@
 class player {
     private:
         std::string* username;
-        int* money;
+        double* money;
     public:
         player();
         ~player();
 
         std::string getUsername()const;
-        int getMoney() const;
+        double getMoney() const;
         void registerUser(const std::string& newUsername, int startingMoney);
-        void setMoney(int newMoney);
-        void subMoney(int amount);
-        void addMoney(int amount);
+        void setMoney(double newMoney);
+        void subMoney(double amount);
+        void addMoney(double amount);
         void saveMoney();
 };
